Inspector V2 Physics properties (#16800)
- [x] Motion type
- [x] Prestep type
- [x] Mass
- [x] Linear/angular damping edit
- [x] Linear/angular velocity value

This PR contains basic support for physics body properties (from a
TransformNode.physicsBody). If there is no physicsBody, right now it
just uses a placeholder line component. Later, we are thinking we can
provide some info and maybe a link to docs.

I (@ryantrem) also included some additional changes/fixes in this PR:
- "Flatten" AccordionPane. There is really no reason at this layer to
have an array of content, since the original purpose of that in the
context of PropertiesPane was to make it so the delegate doesn't have to
be duplicated, but at this layer there is no delegate.
- Update `useProperty` to return undefined if `target` is undefined,
which is more consistent with getting undefined when the property does
not exist.
- Update `DropdownProps.options` to be a readonly array (otherwise there
is a compile error if you try to pass in a readonly array because it
thinks the Dropdown needs to be able to mutate the array).
- Update `Dropdown` to use `useMemo` instead of `useState` so it updates
with props changes.

---------

Co-authored-by: Ryan Tremblay <ryan.tremblay@microsoft.com>