Inspector v2: Rendering Pipelines in Scene Explorer (#16864)
This change is mostly just adding basic support for rendering pipelines
in scene explorer, but required a few additional changes:
- Added a `uniqueId` to rendering pipelines (everything in the tree
needs a unique id, and it is simplest if it is defined directly on the
entity, which is usually the case).
- Added observables for adding/removing rendering pipelines.
- Updated the common properties to handle the case where the `name`
property is readonly (as is the case with rendering pipelines).