Inspector v2: Add missing add/remove observables to scene and use ISceneContext for scene explorer sections (#16869)
This PR contains a couple of related changes:
1. Update the particle system scene explorer section such that it
determines if an entity is a particle system by checking if it is in
scene.particleSystems. This works regardless of the concrete
implementation of IParticleSystem.
2. Doing this though requires the scene. Originally I had scene explorer
pass the scene to the sections various functions to try to make things
simpler, but this is kind of just guessing what the scene explorer
sections will need to do what they want. Since this isn't always the
case (as seen in bullet 1), I decided to just remove this and let the
scene explorer sections get the scene from the ISceneContext if they
need it, and then it is implicitly available in any of the section
functions that need it. This also makes the scene explorer sections more
flexible to put whatever they want, which lets us make scene explorer
more broad if we want (such as including audio, etc.).
3. Added scene observables for adding/removing particle systems.
4. Added scene observables for adding/removing other scene entities that
are not currently available.
5. Added a uniqueId to `IParticleSystem`. This is technically a breaking
change, but I suspect this is ok as there probably aren't many (if any)
custom IParticleSystem implementations. Also we already did this for
ISpriteManager.
6. Added a uniqueId to `FrameGraph`.