Inspector v2: Consolidate light service definitions into one (#16906)
This change mostly just consolidates the different service definitions
for light related properties into a single service definition (all or
none). This just makes it easier to manage, since they are all small
registrations that just associate a specific light properties react
component with a section and an associated predicate. This is similar to
what I did for cameras, and I think this is easier to understand and
maintain.

Also, adding back lost properties for Bone (just the link to the
associated node), which got lost in a smaller consolidation that was
similar to this one.