native morph workaround (#16910)
morphTargetCount is a uniform int;
but bgfx only support vec4 as uniform.
when doing the replacement, morphTargetCount.x is used and comparison
between `int i` and `morphTargetCount.x` which is a float break shader
compilation with GLSL.
casting `morphTargetCount` to int doesn't work: casting qualifier is
removed becaused it's still an int before shader traversers.
Setting morphTargetCount a float in the engine and casting iterators to
float seems to work better.
I've also done the change for WGSL to pass the tests.

Long term is to add (or change) traversers for float/int comparison.