Make CustomShaderBlock resilient to incompatible default values, simplify log/warning/error handling in SFE (#16931)
This change adds type checking for default values when loading a custom
shader block, so you now get an actionable warning in SFE if you get the
type wrong on a custom block you're trying to load. This change also
adds the ability for warnings, not just logs and errors, to show in the
SFE UI. It also simplifies the way logs/warning/errors are recorded in
SFE, no longer requiring an observable to be passed around, but instead
in one place wiring it up to the Logger class, then using Logger
everywhere. This means SF core logs can show up in the SFE UI as well,
such as these new custom shader block warnings this PR adds. This change
also adds watching assets (e.g. scss files) to the SFE VSCode launch
config.