Inspector v2: Improve scene explorer command infra (#16971)
This PR mainly is about improving the scene explorer command
infrastructure in preparation for fleshing out the commands.

- Rework commands so that more state change can be observed. For
example, if you toggle the visiblity of a mesh in the property pane, the
command can reflect the state immediately.
- As was already the case, command buttons on tree view items are only
shown when the item is focused or has pointer hover. However, now toggle
commands that are toggled on will always show so you can easily see the
nodes where you've toggled on a command (such as toggling visibility to
false (not visible)).


https://github.com/user-attachments/assets/c8032108-d988-4523-b6ac-3659499f81ae