Update control.ts to avoid logging to console with text resize logic
Control.ts expects less than 3 layouts to occur but textBlock.ts will layout 3 times when resizeToFit is true -- this is because isDirty is set to true causing a recreation of fontOffset (even if font offset has the proper height) and thus fontOffset.height is getting recalculated to the old value.  We should fix the logic so that fontOffset retains the updated height value. Until then, upping the layout cycle detection error threshold