Add check for build errors before running CopyMinToMax Webpack plugin (#17002)
Previously if UMD failed the webpackPlugin would still try to copy
minified.js file over (which wouldnt exist) so it would error out and
that error would appear in build, hiding the actual underlying error.

This change skips the webpackPlugin copy if there is a build error, and
logs to console to notify.

Co-authored-by: Georgina <gehalper@microsoft.com>