Adding lottie playback using Babylon (#16990)
This PR adds support for Babylon to play a subset of lottie animations.

- These animations will be played on an offscreen canvas and a worker
thread.
- The code takes a dependency only on ThinEngine and as few features as
possible from Babylon to minimize package size.
- Only certain features of lottie are supported currently.
- This is a highly experimental feature, use it at your own risk :)

To use this feature, instantiate an object of the class LottiePlayer and
just set it to play the animation you want:

```
const player = new LottiePlayer(divHtmlElement, 'urltolottie.json');
player.playAnimation();
```

The player will create a canvas and set it to offscreen inside your div
HTML element. By default, the animation will just play once, but you can
control multiple parameters of the playback and the rendering by using
the parameter of playAnimation, which takes an object of type
AnimationConfiguration. Below is the default configuration that will be
merged with the user configuration.

```
/**
 * Default configuration for lottie animations playback.
 */
const DefaultConfiguration: AnimationConfiguration = {
    loopAnimation: false, // By default do not loop animations
    spriteAtlasSize: 2048, // Size of the texture atlas
    gapSize: 5, // Gap around the sprites in the atlas
    spritesCapacity: 64, // Maximum number of sprites the renderer can handle at once
    backgroundColor: { r: 1, g: 1, b: 1, a: 1 }, // Background color for the animation canvas
    scaleMultiplier: 5, // Minimum scale factor to prevent too small sprites,
    devicePixelRatio: 1, // Scale factor,
    easingSteps: 4, // Number of steps to sample easing functions for animations - Less than 4 causes issues with some interpolations
    ignoreOpacityAnimations: true, // Whether to ignore opacity animations for performance
    supportDeviceLost: false, // Whether to support device lost events for WebGL contexts,
};
```