Updating gitignore for devhost public folder (#17007)
When using devhost it is useful to put data in the public folder to test
(like textures or other data), but this should not be checked on the
repo because it's something just useful for local testing. Right now to
avoid checking the data you have to copy them to public and then delete
them by hand, which is easy but annoying.

Adding some lines to gitignore to allow this scenario.