FrameGraph: adds support for frozen meshes + allows no output color texture for the geometry buffer renderer task (#17040)
This PR adds support for frozen meshes to frame graphs.

Note that we have a bug when using multiple active cameras in a scene
with meshes + instances. If we freeze the scene, only some instances
(those that were culled and displayed by the last camera in the list)
are displayed, due to the way we currently handle frozen meshes (we have
a single `scene._activeMeshes` list, which is populated by a call to
`_evaluateActiveMeshes` when `freezeActiveMeshes` is called). I'm not
sure how to fix this bug in the current code, but the frame graph
version of the same example works correctly because it uses object
renderers to render meshes:
* without frame graph: https://playground.babylonjs.com/#9SDNZF#30
* with frame graph: https://playground.babylonjs.com/#9SDNZF#29

The meshes are frozen after two seconds. In the first PG, some instances
disappear along with the plane because they are displayed by the first
camera. Note that you must wait for the PR to be merged to test the
second PG, because otherwise the meshes will not be frozen even after
calling `scene.freezeActiveMeshes()`!

In addition, the PR allows you to not define any output color textures
for the geometry buffer renderer. This allows you to generate only a
depth render of the scene and reuse the depth buffer when rendering the
scene later in the graph.