glTF Exporter: Use stored image data of Textures (#17044)
This PR introduces an optimization to the glTF exporter. When possible,
the exporter will now leverage the original URL or `_buffer` of a
`Texture` instead of re-encoding the image data. This can lead to
smaller export size, faster export times, and less loss on roundtrips,
especially when the original image format was already compressed.

This optimization is applied if the following conditions are met:

- The texture's source is URL-based. We assume that if a texture was
loaded from a URL, its pixel data has not been modified.
- The `engine.doNotHandleContextLost` setting is `false` (or, if `true`,
that `Texture.url` is able to be used alone.)
- The texture was not inverted on the Y-axis during import.

If these conditions are not met, the exporter will fall back to the
previous behavior of reading the pixels from the GPU and re-encoding
them based on `Texture.mimeType`.

<details>
<summary>
A Note on Texture.mimeType...
</summary>

There is some ambiguity surrounding the `Texture.mimeType` property. It
has sometimes been used to influence the glTF export type, for example,
to preserve an original image's MIME type when it was not implicitly
known. While this is useful, it also had the side effect of allowing
users to re-encode textures on export by setting a different `mimeType`
(e.g., from "image/png" to "image/webp").

This optimization may break this usage. Going forward,
`Texture.mimeType` should be considered a hint about the underlying
data's format, not a directive to encode to a new format. A separate,
more explicit mechanism for format conversion should be introduced in
the future.
</details>