Add Lazy helper class and use it for dynamic imports (#17045)
This is a follow up from
https://github.com/BabylonJS/Babylon.js/pull/17043. It adds a generic
`Lazy<T>` helper class, and then I use it in gltfLoader. There are a
couple nice things about this pattern:

1. It keeps the import paths all in one place at the top of the file,
rather than sprinkled throughout the file, which makes it easier to
understand dependencies quickly.
2. It reduces the amount of code and makes it trivial to use this
pattern anywhere we have dynamic imports in high traffic critical parts
of the code.