Inspector v2: Rework BoundProperty to have no conditional hooks (#17041)
This changes BoundProperty to remove the conditional hooks (see
https://react.dev/reference/rules/rules-of-hooks). The strategy here is
to first do a useProperty in the "outer" component so we can observe the
value of that property to determine if we really ideally should be using
a more specific property hook that watches nested properties like x/y/z
of a Vector3 for example. The desired hook is then used as the
dependency for a useMemo that creates and returns an "inner" component
that uses the desired hook. This eliminates conditional hooks.