Inspector v2: Picking toolbar (#17058)
This PR adds a picking button to the toolbar for picking objects from
the scene.

<img width="406" height="189" alt="image"
src="https://github.com/user-attachments/assets/324bdb17-6340-410e-9ebc-d2eedc73c2ea"
/>

The picking logic is similar to inspector v1, but fixes a handful of
bugs as well.
- Picking gizmos doesn't work in v1, at least not in any way I expect it
to as a user or looking at the existing code. It works in v2.
- Picking meshes with morph target or skeletal animations does not work
in v1. It works in v2.

The picking logic also uses the "Ignore Backfaces for Picking" setting
that @deltakosh already added to the settings pane.

<img width="396" height="161" alt="image"
src="https://github.com/user-attachments/assets/ffbd1542-1099-4705-8734-1fdf64b20957"
/>

Other changes:
- Fixed a bug that was [introduced 5 years
ago](https://github.com/BabylonJS/Babylon.js/pull/8853/files) where the
`isHovered` state on a gizmo is only updated on pointer click, not on
pointer move, which means if you click on it, it just stays in the
"hovered" state indefinitely.