Pass through gizmo drag observer messages (#17055)
I was recently working on a feature that utilized the PositionGizmo,
adjusting its behavior based on whether the shift key was pressed. I was
surprised to find that the pointer info wasn't available at the
PositionGizmo level because the notification packet from the
PointerDragBehavior isn't being passed through. This seems to have been
done intentionally, though I'm not sure why.

This PR proposes that the notification packets for the specialized gizmo
PointerDragBehavior be passed through their onDragObservables. This
implementation keeps the Observer type dynamic to ease future
development.

If the current behavior was intentional, I'm curious to know why.