Viewer: Fix perf issue with hotspots (#17076)
This change mainly switches from using getVerticesData to
getVertexBuffer+EnumerateFloatData to drastically reduce the cost of
querying hotspots (about a 266x improvement from my measurements). The
issue with getVerticesData is that it makes copies of large buffers, and
we were doing this three times per hotspot (one for each point of the
triangle). On the other hand, getVertexBuffer just accesses the existing
buffer without copying, and EnumerateFloatData basically gives us an
easy way to index into that buffer to get just the data subset
corresponding to a single element.

So previously it only took about 15 hotspots to use up the entire 16ms
frame time on my desktop machine. With the new approach, it takes about
60 microseconds to query 15 hotspots. If needed, we could possibly
further improve this in the future by adding caching, but if this is
fast enough then it is also more robust since it would work correctly
when the hotspot definition changes (e.g. indices for the position
buffer) or the position buffers themselves.