Adding support for the player to take a json object (#17078)
The current API of the player only accepted the URL of a file to load.
This change also adds supporting a JSON object as the input instead of
the URL, as in certain scenarios the object may have already been
parsed, and we don't need to parse it ourselves.

Also added support for the devHost to test this scenario with the QSP
useUrl=boolean. This QSP will tell the devhost to either pass the URL to
the player or to download the file and parse the content and pass the
json instead. Default is using the URL.