Improve Smart Filter optimizer to handle more cases (#17077)
With this change, the Smart Filter optimizer now handles these
scenarios:
- Helper functions (functions other than the main function) can now read
uniforms, although if a function does, a copy of it will be emitted for
each instance of the block that is optimized into a single shader
- Helper functions can now read consts and defines
- Helpers can now call other helpers regardless of the order in which
they are defined

The key change is that renames are not done immediately (and only in the
main function) but are accumulated and then applied to the main function
as well as all the helper functions at the end of optimizing a block.

This also fixes a bug in the scenario where multiple instances of a
block are collapsed into a single optimized block, and one of the
instances reads a texture that comes from a non-optimizable block. In
that scenario, the optimizer will create an input connection point to
connect to that non-optimizable block, but in some situations, it would
be wired up to the wrong instance of the block.